+++ chrome/devtools.js
@@ -1,0 +1,8 @@
+chrome.devtools.panels.create(
+  "XnlReveal",
+  "images/icon16.png",
+  "panel.html",
+  function (panel) {
+    // code invoked once the panel is created
+  }
+);