+++ firefox/devtools.js
@@ -1,0 +1,7 @@
+browser.devtools.panels.create(
+  "XnlReveal",
+  "images/icon16.png",
+  "panel.html"
+).then((panel) => {
+  // panel loaded
+});